move(/service): ReservationService 를 /service 이하로 이동
