+++ src/main/java/roomescape/controller/ReservationController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
-import roomescape.ReservationService;
+import roomescape.service.ReservationService;
 import roomescape.controller.dto.ReservationRegisterDto;
 import roomescape.controller.dto.ReservationResponseDto;
