+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,6 +1,8 @@
 package roomescape.controller;
 
 import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 import roomescape.controller.dto.ReservationTimeRegisterDto;
 import roomescape.controller.dto.ReservationTimeResponseDto;
 import roomescape.domain.ReservationTime;
@@ -43,7 +46,12 @@ public class ReservationTimeController {
 
     @ResponseBody
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable("id") final Long id) {
-        reservationTimeRepository.deleteById(id);
+    public ResponseEntity<String> deleteById(@PathVariable("id") final Long id) {
+        try {
+            reservationTimeRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } catch (ResponseStatusException e) {
+            return new ResponseEntity<>(e.getMessage(), e.getStatusCode());
+        }
     }
 }
