+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.service;
 
 import java.util.List;
 import java.util.Optional;
